Přidání abstraktních kontrolerů pro AJAX a synchronní požadavky
Všechny AJAX kontrolery budou dědit z nové abstraktní třídy AjaxController
Všechny synchronní kontrolery (zobrazující nějaký pohled) budou dědit z nové abstraktní třídy SynchronousController
Obě tyto abstraktní třídy poté dědí z abstraktní třídy Controller
Všechny neabstraktní metody jsou přesunuty z Controller.php do SynchronousController.php, protože je téměř výhradně využívají kotnrolery nastavující nějaký pohled a ve zbylých případech bude kód předělán
Ostatní kontrolery zatím nemají aktualizované reference a tak žádný z nich nefunguje
